feat: Add support for `--find_links`
Any package included at the `find_links` URI will be collected from that
location, rather than querying the `index_url` for it. It's worth noting
that this behaves differently from `pip`s interpretation of
`--find-links` which simply add an extra source to package resolution,
from which `pip` will find the best candidate for package sources. But
this is a much simpler change which allows bundling of local package
files.